+++ src/ui/event/key/event.rs
@@ -1,37 +1,33 @@
 // devela::ui::event::key::event
 //
-//! Defines [`EventKey`], [`EventKeyFfi`], [`KeyState`].
+//! Defines [`EventKey`], [`EventKeyFfi`].
 //
-// TOC
-// - struct EventKey
-// - struct KeyState
-// - ffi
-//   - struct EventKeyFfi
-//   - impls
 
-#[cfg(all(feature = "js", not(windows)))]
-use crate::WebEventKind;
-use crate::{ConstInit, EventTimestamp, Key, KeyMods};
+use crate::{ConstInit, EventTimestamp, Key, KeyMods, KeyState};
 
 /// Represents a keyboard event.
 ///
 #[doc = "See also [`EventKeyFfi`]."]
 #[derive(Clone, Copy, Debug, Default, PartialEq, Eq, Hash)]
 pub struct EventKey {
+    /// The time stamp of when the event occurred.
+    pub timestamp: Option<EventTimestamp>,
+
     /// The key representing the human-readable code.
     ///
     /// This corresponds to X11's keysym.
     pub semantic: Key,
+
     /// The key representing the hardware scan code.
     ///
     /// This corresponds to X11's scancode XKB mapped.
     pub physical: Key,
-    /// The state of the key (pressed or released).
-    pub state: KeyState,
+
     /// The active modifiers of the key (e.g., Shift, Ctrl).
     pub mods: KeyMods,
-    /// The time stamp of when the event occurred.
-    pub timestamp: Option<EventTimestamp>,
+
+    /// The state of the key (pressed or released).
+    pub state: KeyState,
 }
 impl ConstInit for EventKey {
     const INIT: Self = Self {
@@ -43,25 +39,14 @@ impl ConstInit for EventKey {
     };
 }
 
-/// Represents the state of a [`Key`].
-#[repr(u8)]
-#[derive(Clone, Copy, Debug, Default, PartialEq, Eq, Hash)]
-pub enum KeyState {
-    /// The key was pressed.
-    #[default]
-    Press,
-    /// The key was released.
-    Release,
-}
-impl ConstInit for KeyState {
-    const INIT: Self = Self::Press;
-}
-
 #[cfg(ffi··)]
+pub use ffi::*;
+#[cfg(ffi··)]
+#[rustfmt::skip]
 #[cfg_attr(nightly_doc, doc(cfg(ffi··)))]
 mod ffi {
     use super::*;
-    use crate::{ConstInit, KeyFfi, f32bits};
+    use crate::{ConstInit, KeyFfi, f32bits, is};
 
     #[doc = crate::_TAG_FFI!()]
     /// An FFI-safe version of [`EventKey`].
@@ -72,13 +57,17 @@ mod ffi {
         #[doc = crate::_TAG_FFI!()]
         /// The key representing the human-readable code.
         pub semantic: KeyFfi,
+
         #[doc = crate::_TAG_FFI!()]
         /// The key representing the hardware scan code.
         pub physical: KeyFfi,
+
         /// The state of the key (pressed or released).
         pub state: KeyState,
+
         /// The active modifiers of the key (e.g., Shift, Ctrl).
         pub mods: KeyMods,
+
         #[doc = crate::_TAG_FFI!()]
         /// The time stamp of when the event occurred.
         pub timestamp: f32bits,
@@ -92,11 +81,7 @@ mod ffi {
                 physical: self.physical.to_ffi(),
                 state: self.state,
                 mods: self.mods,
-                timestamp: if let Some(t) = self.timestamp {
-                    t.get_non_niche()
-                } else {
-                    f32bits::INIT
-                },
+                timestamp: is![let Some(t) = self.timestamp; t.get_non_niche(); f32bits::INIT],
             }
         }
         /// Converts `EventKeyFfi` to `EventKey`.
@@ -110,34 +95,8 @@ mod ffi {
             }
         }
     }
-    crate::items! {
-        impl From<&EventKey> for EventKeyFfi { fn from(e: &EventKey) -> Self { EventKey::to_ffi(e) } }
-        impl From<&EventKeyFfi> for EventKey { fn from(e: &EventKeyFfi) -> Self { Self::from_ffi(e) } }
-        impl From<EventKey> for EventKeyFfi { fn from(e: EventKey) -> Self { EventKey::to_ffi(&e) } }
-        impl From<EventKeyFfi> for EventKey { fn from(e: EventKeyFfi) -> Self { Self::from_ffi(&e) } }
-    }
-
-    #[cfg(all(feature = "js", not(windows)))]
-    #[cfg_attr(nightly_doc, doc(cfg(feature = "js")))]
-    impl KeyState {
-        /// Converts a `WebEventKind` to `KeyState`, if applicable.
-        #[must_use]
-        pub const fn from_js(from: WebEventKind) -> Option<KeyState> {
-            match from {
-                WebEventKind::KeyDown => Some(KeyState::Press),
-                WebEventKind::KeyUp => Some(KeyState::Release),
-                _ => None,
-            }
-        }
-        /// Converts a `KeyState` to `WebEventKind`.
-        #[must_use]
-        pub const fn to_js(self) -> WebEventKind {
-            match self {
-                KeyState::Press => WebEventKind::KeyDown,
-                KeyState::Release => WebEventKind::KeyUp,
-            }
-        }
-    }
+    impl From<&EventKey> for EventKeyFfi { fn from(e: &EventKey) -> Self { EventKey::to_ffi(e) } }
+    impl From<&EventKeyFfi> for EventKey { fn from(e: &EventKeyFfi) -> Self { Self::from_ffi(e) } }
+    impl From<EventKey> for EventKeyFfi { fn from(e: EventKey) -> Self { EventKey::to_ffi(&e) } }
+    impl From<EventKeyFfi> for EventKey { fn from(e: EventKeyFfi) -> Self { Self::from_ffi(&e) } }
 }
-#[cfg(ffi··)]
-pub use ffi::*;