+++ src/ui/event/mod.rs
@@ -9,6 +9,9 @@ mod pointer; // Event[Button[State]|Mouse|Pointer[Type]|Wheel]
 mod time; // EventTimestamp
 mod window; // EventWindow
 
+#[cfg(test)]
+mod tests;
+
 crate::structural_mods! { // _mods
     _mods {
         pub use super::{